requirements.txt
moviepy==1.0.3
openai_whisper==20230314
